// frontend/app/models/kupovina.ts
import { Korisnik } from "./korisnik.model";

export class kupovinaNekretnina {
    opis: string;
    adresa: string;
    tip: string;
    sprat: string;
    kvadratura: string;
    sobe: string;
    namestenost: string;
    izdavanje: string;
    galerija: string[];
    // cena: string;
    cena:number;
    vlasnik: string;
    id: string;
    promo:boolean;
    kupac:string;
}